keep all errors in one module
