+++ src/firebolt/client/__init__.py
@@ -1,7 +1,7 @@
 from firebolt.client.client import (
     DEFAULT_API_URL,
-    AuthenticationError,
     FireboltAuth,
     FireboltClient,
 )
 from firebolt.client.singleton import get_firebolt_client, init_firebolt_client
+from firebolt.common.exception import AuthenticationError
